requirements.txt
aiocontextvars==0.2.2
asgiref==3.4.1
contextvars==2.4
dataclasses==0.8
Deprecated==1.2.14
Django==3.2.25
immutables==0.19
opentelemetry-api==1.12.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-django==0.33b0
opentelemetry-instrumentation-wsgi==0.33b0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.1.1
wrapt==1.16.0
